add search to user profile
